Fix issue with role composition onto subclasses.
Metaclass re-initialization was happening too late - i.e. after the methods
from the roles were composed, but before the method modifiers from the roles.

This resulted in the metaclass reinitialization losing the methods from the
roles, and then trying to apply the modifiers wrapping the lost methods
which, of course, entirely failed.

The solution is that we don't actaully need to perform force reinitialization
at this point, and so removing that line fixes everything.